Fix erlang clause error in non-grant case
There was a gap in testing coverage and in the case that the code path
was being followed for a check that was a non-grant acl it would result
in an erlang error due to an ok being returned when a boolean was
expected. This change fixes the error and adds tests for it.